refactor: クラス名をMainからAppに変更し、関連するインポートやメトリクスファクトリーを追加してアプリケーションの構成を整理する
